+++ demo/component/Common.js
@@ -1,0 +1,11 @@
+import Demobox from './Demobox';
+import ExampleHeader from './ExampleHeader';
+import ObjectView from './ObjectView';
+import PropsTable from './PropsTable';
+
+export {
+  Demobox,
+  ExampleHeader,
+  ObjectView,
+  PropsTable
+};