+++ demo/routes/Icon/components/Icon.js
@@ -1,15 +1,12 @@
 import React, { Component } from 'react';
 import _map from 'lodash/map';
-import {
-  Flex,
-  Card,
-  Icon,
-  Sheet
-} from '../../../../src';
+import { Flex, Card, Icon } from 'react-timo';
+import {ExampleHeader, PropsTable, Demobox} from 'demo-component';
 
 import icons from '../config/icons';
-import props from '../config/props';
-import ExampleHeader from '../../../component/ExampleHeader';
+import iconProps from '!!docgen-loader!../../../../src/component/Icon';
+import code from '!!raw-loader!./Demo';
+import Demo from './Demo';
 
 class IconExample extends Component {
   constructor(props) {
@@ -23,18 +20,10 @@ class IconExample extends Component {
           title="Icon Component"
           desc="Material Design Iconic Font is a full suite of official material design icons (created and maintained by Google), with additional community-designed icons and brands icons for easy scalable vector graphics on websites or desktop"
         />
-        <Card title="Props">
-          <div className="row">
-            <div className="col-sm-12">
-              <Sheet list={props}>
-                <Sheet.SheetColumn field="props" name="Props名称"/>
-                <Sheet.SheetColumn field="isRequired" name="是否必须"/>
-                <Sheet.SheetColumn field="default" name="默认值"/>
-                <Sheet.SheetColumn field="comment" name="Props说明"/>
-              </Sheet>
-            </div>
-          </div>
-        </Card>
+        <Demobox code={code}>
+          <Demo />
+        </Demobox>
+        <PropsTable list={iconProps}/>
         {
           _map(icons, (value, key) => {
             return (