+++ src/component/Collapse/index.js
@@ -1,0 +1,4 @@
+import Collapse from './Collapse';
+import './_collapse.scss';
+
+export default Collapse;