+++ src/component/Highlight/index.js
@@ -1,0 +1,4 @@
+import Highlight from 'react-highlight';
+import 'highlight.js/styles/arduino-light.css';
+
+export default Highlight;