+++ src/component/Icon/index.js
@@ -7,6 +7,7 @@
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import cx from 'classnames';
 import _noop from 'lodash/noop';
 import _omit from 'lodash/omit';
 
@@ -22,14 +23,31 @@ import _omit from 'lodash/omit';
 
 class Icon extends Component {
   static propTypes = {
+    /**
+     * Icon 类型
+     */
     type: PropTypes.string.isRequired,
+    /**
+     * Icon组件额外的内联样式
+     */
     style: PropTypes.object,
+    /**
+     * Icon 自动对齐 .zmdi-hc-fw
+     */
+    autoFixed: PropTypes.bool,
+    /**
+     * Icon组件额外的类名
+     */
     className: PropTypes.string,
+    /**
+     * Icon组件触发点击事件回调函数
+     */
     onClick: PropTypes.func
   }
   static defaultProps = {
     style: {},
     className: '',
+    autoFixed: false,
     onClick: _noop
   }
   constructor(props) {
@@ -47,11 +65,15 @@ class Icon extends Component {
       type,
       style,
       className,
+      autoFixed,
       ...rest
     } = this.props;
+    const cls = cx('zmdi', `zmdi-${type}`, {
+      'zmdi-hc-fw': autoFixed
+    }, className);
     return (
       <i
-        className={`zmdi zmdi-${type} ${className}`}
+        className={cls}
         style={style}
         onClick={this.handleClick}
         {..._omit(rest, ['onClick'])}